requirements.txt
blinker==1.7.0
cffi==1.16.0
click==8.1.7
cryptography==41.0.5
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.3.2
MarkupSafe==2.1.3
numpy==1.26.1
pandas
psycopg2-binary==2.9.9
pycparser==2.21
PyJWT==2.8.0
python-dateutil==2.8.2
pytz==2023.3.post1
scikit-learn
scipy==1.11.3
six==1.16.0
SQLAlchemy==2.0.23
threadpoolctl==3.2.0
typing_extensions==4.8.0
Werkzeug==3.0.1
gunicorn==19.9.0
